Some changes

+++ app/src/main/java/com/pointlessapps/amnesia/model/CategoryModel.kt
@@ -9,5 +9,5 @@ import java.util.*
 data class CategoryModel(
     val id: Long = UUID.randomUUID().mostSignificantBits,
     val name: String,
-    @ColorInt val color: Int,
+    @ColorInt val color: ULong,
 ) : Parcelable
