+++ domain/src/main/java/com/pointlessapps/amnesia/domain/notes/dto/Category.kt
@@ -7,5 +7,5 @@ import androidx.annotation.Keep
 data class Category(
     val id: Long,
     val name: String,
-    @ColorInt val color: Int,
+    @ColorInt val color: ULong,
 )
